Refactor the listing importer script
