+++ backend/core/scripts/minimal-listing.json
@@ -1,0 +1,29 @@
+{
+  "name": "Minimal Listing",
+  "status": "active",
+  "CSVFormattingType": "basic",
+  "countyCode": "Alameda",
+  "preferences": [],
+  "buildingAddress": {
+    "city": "Oakland",
+    "state": "CA",
+    "street": "Main St",
+    "zipCode": "94501",
+    "latitude": "37.80",
+    "longitude": "122.27"
+  },
+  "units": [
+    {
+      "unitType": {
+        "name": "oneBdrm",
+        "numBedrooms": 1
+      },
+      "status": "available"
+    }
+  ],
+  "assets": [],
+  "applicationMethods": [],
+  "events": [],
+  "displayWaitlistSize": false,
+  "jurisdictionName": "Alameda"
+}
